Update page.js
This changes the direction in which the smurf is pointing. Rest assured that this code works when I tested it out on my computer, and considering that every direction has a value to it now, what we can do is use that value to determine the direction that the bullet is fired.